Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -1,10 +1,10 @@
 #include <stdio.h>
 
 int main() {
-    int r;
+    float r;
     float area;
     scanf("%d", &r);
-    area = 3.14*r*r;
+    area = 3.14 * r * r;
     printf("Area: %.2f", area);
     return 0;
 }
